Footer改修 責務を明確に
