+++ src/components/Layout/Footer/index.ts
@@ -1,0 +1,1 @@
+export { Footer } from './internal/Footer';
