+++ src/components/Layout/Footer/internal/Footer/index.ts
@@ -1,0 +1,1 @@
+export { default as Footer } from './internal/Footer.component';
