+++ src/components/Layout/Footer/internal/FooterIcons/index.ts
@@ -1,0 +1,2 @@
+export { default as FooterIcon } from './internal/FooterIcon.component';
+export type { FooterIconsProps } from './internal/FooterIcon.types';
