+++ src/components/Layout/Footer/internal/FooterIcons/internal/FooterIcon.types.ts
@@ -1,0 +1,7 @@
+import { FooterType } from '../../types';
+
+export type FooterIconsProps = {
+  type: FooterType;
+  isSelected: boolean;
+  onClick: () => void;
+};
