+++ src/components/Layout/Footer/internal/types.ts
@@ -1,0 +1,1 @@
+export type FooterType = 'Map' | 'Qr' | 'Search';
